requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.3
idna==3.6
platformdirs==4.2.0
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
requests-file==2.0.0
sqlparse==0.4.4
tldextract==5.1.1
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
zope.event==5.0
zope.interface==6.1
